+++ testkit/src/main/scala/akka/kafka/testkit/javadsl/EmbeddedKafkaTest.scala
@@ -3,7 +3,7 @@
  * Copyright (C) 2016 - 2018 Lightbend Inc. <http://www.lightbend.com>
  */
 
-package akka.kafka.javadsl
+package akka.kafka.testkit.javadsl
 
 import net.manub.embeddedkafka.{EmbeddedKafka, EmbeddedKafkaConfig}
