+++ tests/src/test/scala/akka/kafka/scaladsl/SpecBase.scala
@@ -5,6 +5,7 @@
 
 package akka.kafka.scaladsl
 
+import akka.kafka.testkit.scaladsl.{EmbeddedKafkaLike, ScalatestKafkaSpec}
 import org.scalatest.concurrent.{Eventually, ScalaFutures}
 import org.scalatest.{Matchers, WordSpecLike}
